font table interpreter: moved d stuff responsible for font table into d2fileformats/d2font
