+++ d2core/d2gui/layout.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"image/color"
 
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2fileformats/d2font"
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2interface"
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2math"
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2util"
@@ -532,7 +533,7 @@ func (l *Layout) createButton(renderer d2interface.Renderer, text string,
 	return button, nil
 }
 
-func (l *Layout) loadFont(fontStyle FontStyle) (*d2asset.Font, error) {
+func (l *Layout) loadFont(fontStyle FontStyle) (*d2font.Font, error) {
 	config := getFontStyleConfig(fontStyle)
 	if config == nil {
 		return nil, errors.New("invalid font style")
