+++ d2core/d2ui/label.go
@@ -5,7 +5,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/OpenDiablo2/OpenDiablo2/d2core/d2asset"
+	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2fileformats/d2font"
 
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2interface"
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2util"
@@ -19,7 +19,7 @@ type Label struct {
 	*BaseWidget
 	text            string
 	Alignment       HorizontalAlign
-	font            *d2asset.Font
+	font            *d2font.Font
 	Color           map[int]color.Color
 	backgroundColor color.Color
